requirements.txt
pandas==1.5.3
numpy==1.24.3
pytz==2023.3
PyJWT==2.8.0
pytest
boto3